reflect: ToStringWithFallback
